+++ python/docker/run_dictation_client_python.sh
@@ -135,4 +135,4 @@ docker run --rm -it \
 --group-add audio \
 --network host \
 "${docker_image}" \
-/dictation_client_python/run.sh "${opts[@]}"
+/dictation-client/python/run.sh "${opts[@]}"